requirements.txt
asgiref==3.7.2
attrs==22.2.0
Brlapi==0.8.5
cffi==1.15.1
cryptography==41.0.3
distlib==0.3.7
Django==4.2.5
filelock==3.12.2
lensfun==0.3.4
louis==3.27.0
meson==1.2.1
netsnmp-python==1.0a1
packaging==23.1
Pillow==10.0.1
platformdirs==3.10.0
ply==3.11
pwquality==1.4.5
pycairo==1.24.0
pycparser==2.21
pygame==2.4.0
PyGObject==3.46.0
pytesseract==0.3.12
pytz==2023.3.post1
pyxdg==0.28
setproctitle==1.3.2
sqlparse==0.4.3
tqdm==4.66.1
typing_extensions==4.8.0
virtualenv==20.24.3
